requirements.txt
chainlit
google-generativeai
langchain
Pillow 
SpeechRecognition